SMS-Back-End/apigateway/requirements.txt
# This requirements file lists all third-party dependencies for this project.
# Este fichero de requerimientos lista todas las dependencias de terceros para este proyecto.
#
# Ejecute 'pip install -r requirements.txt -t lib/' para instalar estas dependencias
# en el subdirectorio `lib/` .
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.

GoogleAppEngineCloudStorageClient
